+++ client/src/components/HeaderOption.css
@@ -2,7 +2,8 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    margin-right: 20px;
+    margin-right: 10px;
+    margin-left: 10px;
     color: gray;
     cursor: pointer;
 }
@@ -15,5 +16,5 @@
 
 .headerOption:hover {
     color: black;
-    
+
 }
