+++ client/src/pages/account/Profile.jsx
@@ -1,10 +1,7 @@
 import { useEffect } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { useCookies } from "react-cookie";
+import { Link } from "react-router-dom";
 import { useSelector, useDispatch } from "react-redux";
 import { selectCurrentUser } from "../../store/auth/userSlice";
-import { useLogoutMutation } from "../../store/auth/authApiSlice";
-import { logOut } from "../../store/auth/userSlice";
 import { useGetCompetitionsQuery } from "../../store/game/competitionApiSlice";
 import { setCompetitions } from "../../store/game/competitionSlice";
 import Sidebar from "./Sidebar";
@@ -14,14 +11,7 @@ import "./Profile.css";
 
 function Profile() {
   const user = useSelector(selectCurrentUser);
-  const token = user.token;
-
-  const [, removeCookie] = useCookies(["token"]);
-  const navigate = useNavigate();
   const dispatch = useDispatch();
-
-  const [logout] = useLogoutMutation();
-
   const { data: competitions, isLoading: isCompsLoading, refetch: refetchComps } = useGetCompetitionsQuery(user.id);
 
   useEffect(() => {
@@ -49,18 +39,6 @@ function Profile() {
     );
   });
 
-  async function handleLogout() {
-    try {
-      await logout({ headers: { Authorization: `Bearer  ${token}` } });
-      dispatch(logOut());
-      removeCookie("token");
-      localStorage.clear();
-      navigate("/");
-    } catch (error) {
-      console.error("Error logging out", error);
-    }
-  }
-
   const content =
     (<div className="profile__main">
       {/* <Header /> */}
@@ -71,7 +49,6 @@ function Profile() {
           email={user.email}
           gender={user.gender}
           location={user.location}
-          handleLogout={handleLogout}
         />
         <MainFeed
           current_competitions={mapComps}
